refactor: 타입 도메인별 분리
