+++ src/components/statistics/InlineTimeline.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useTheme } from '@/hooks/useTheme';
-import { CycleData, CycleSegment } from '@/types/cycle';
+import { CycleData, CycleSegment } from '@/types/domains/usage/cycle';
 import { ArrowLeft, Clock, Coffee, UserX } from 'lucide-react';
 import React from 'react';
