+++ src/types/common/index.d.ts
@@ -1,0 +1,3 @@
+// 공통 타입들
+export * from './apiResponse';
+export * from './error';
