+++ src/types/domains/index.d.ts
@@ -1,0 +1,6 @@
+// 모든 도메인 타입들
+export * from './user';
+export * from './leaderboard';
+export * from './subscription';
+export * from './usage';
+export * from './payment';
