+++ src/types/domains/leaderboard/index.d.ts
@@ -1,0 +1,2 @@
+// 리더보드 관련 타입들
+export * from './leaderBoard';
