+++ src/types/domains/payment/index.d.ts
@@ -1,0 +1,2 @@
+// 결제 관련 타입들
+export * from './paymentMethod';
