+++ src/types/domains/subscription/index.d.ts
@@ -1,0 +1,2 @@
+// 구독 관련 타입들
+export * from './subscription';
