+++ src/types/domains/user/index.d.ts
@@ -1,0 +1,2 @@
+// 사용자 관련 타입들
+export * from './user';
